==> Stars - Hospitals/04_sas_input.Rmd <==
---
title: "R Notebook"
output:
html_notebook: default
---

```{r}
source("init.R")
library(data.table)

load(file = file.path(dirs$rdata, "all_data_list.RData"))

all_mort_readm_data_df <- all_data_list$all_mort_readm_data_df
all_maternal_data_df   <- all_data_list$all_maternal_data_df
all_hai_data_df        <- all_data_list$all_hai_data_df
all_hcahps_data_df     <- all_data_list$all_hcahps_data_df
all_timely_data_df     <- all_data_list$all_timely_data_df
all_imaging_data_df    <- all_data_list$all_imaging_data_df
all_ipfqr_data_df      <- all_data_list$all_ipfqr_data_df
all_hospital_info_df   <- all_data_list$all_hospital_info_df
all_measure_info_df    <- all_data_list$all_measure_info_df
all_va_hospital_df     <- all_data_list$all_va_hospital_df
all_sas_input_df       <- all_data_list$all_sas_input_df
all_national_rates_df  <- all_data_list$all_national_rates_df
measure_id_history_df  <- all_data_list$measure_id_history_df
```




```{r}
latest_sas_measures_df <-
  all_measure_info_df %>%
  filter(cc_date == "2023-01-01", !is.na(sas_measure_id)) %>%
  #filter(cc_date == "2021-07-01", !is.na(sas_measure_id)) %>%
  select(measure_id_sas = sas_measure_id, measure_id_std = measure_id,
         measure_group, higher_is_better, sas_measure_id, sas_measure_stat)

sas_measure_id_conversion_df <-
  measure_id_history_df %>%
  filter(cc_date >= "2018-01-01") %>%
  pivot_longer(
    -cc_date, names_to = "measure_id_std", values_to = "measure_id_historical"
  ) %>%
  filter(!is.na(measure_id_historical))

sas_measure_df <-
  latest_sas_measures_df %>%
  left_join(sas_measure_id_conversion_df, by = "measure_id_std") %>%
  mutate(
    measure_id_sas = case_when(
      measure_id_sas == "PSI_4_SURG_COMP" ~ "PSI_04",
      measure_id_sas == "PSI_90_SAFETY"   ~ "PSI_90",
      TRUE ~ measure_id_sas
    )
  ) %>%
  arrange(measure_id_std, cc_date)

sas_measure_df %>%
  filter(grepl("IMM", measure_id_sas) | grepl("IMM", measure_id_std))
```



# Measures besides HAI and HCAHPS

```{r}
sas_most_groups_df <-
  bind_rows(
    all_mort_readm_data_df %>%
      filter(cc_date >= "2018-01-01") %>% select(cc_date, data),
    all_maternal_data_df %>%
      filter(cc_date >= "2018-01-01") %>% select(cc_date, data),
    all_timely_data_df %>%
      filter(cc_date >= "2018-01-01") %>% select(cc_date, data),
    all_imaging_data_df %>%
      filter(cc_date >= "2018-01-01") %>% select(cc_date, data)
  ) %>%
  unnest(data) %>%
  filter(!(
    (measure_id == "IMM_3") & (provider_id == "220126") & (is.na(score))
  )) %>%
  #filter(cc_date <= "2019-01-01", grepl("IMM", measure_id)) %>%
  inner_join(
    sas_measure_df,
    by = c("cc_date", "measure_id" = "measure_id_historical")
  ) %>%
  select(
    cc_date,
    provider_id,
    measure_id_sas,
    score,
    DEN
  ) %>%
  pivot_wider(
    id_cols = c(cc_date, provider_id),
    names_from  = measure_id_sas,
    names_glue  = "{measure_id_sas}_{.value}",
    values_from = c(score, DEN)
  )

colnames(sas_most_groups_df) <-
  gsub("_score$", "", colnames(sas_most_groups_df))

sas_most_groups_df %>% select(cc_date, provider_id, IMM_3)


all_measure_info_df %>%
  filter(cc_date >= "2018-01-01", grepl("PSI.*(0?4|90)", measure_id)) %>%
  filter(!is.na(sas_measure_id))

temp_df <- all_timely_data_df %>% filter(cc_date == "2018-01-01")
temp_df$data[[1]] %>% count(measure_id)
sas_measure_df %>% filter(grepl("IMM", measure_id_historical))
```


# HAI

```{r}
sas_hai_df <-
  all_hai_data_df %>%
  filter(cc_date >= "2018-01-01") %>%
  select(cc_date, data) %>%
  unnest(data) %>%
  filter(!is.na(CI_LOWER) | !is.na(CI_UPPER)) %>%
  select(
    cc_date,
    provider_id,
    measure_id,
    SIR,
    DEN_PRED,
    DEN_VOL
  ) %>%
  pivot_wider(id_cols = c(cc_date, provider_id),
              names_from = measure_id,
              names_glue = "{measure_id}_{.value}",
              values_from = c(SIR, DEN_PRED, DEN_VOL))

colnames(sas_hai_df) <- gsub("_SIR$", "", colnames(sas_hai_df))

sas_hai_df
```



# HCAHPS

```{r}
temp_sas_hcahps_df <-
  all_hcahps_data_df %>%
  filter(cc_date >= '2018-01-01') %>%
  select(cc_date, data) %>%
  unnest(data) %>%
  select(
    cc_date, provider_id, measure_id,
    STAR_RATING, TOP_BOX,
    H_NUMB_COMP, H_RESP_RATE_P
  )

hcahps_star_measure_ids <-
  sas_measure_df %>%
  filter(grepl("H[_].*.[_]STAR_RATING", measure_id_sas)) %>%
  pull(measure_id_std) %>% unique

hcahps_star_df <-
  temp_sas_hcahps_df %>%
  filter(measure_id %in% hcahps_star_measure_ids) %>%
  transmute(cc_date, provider_id, STAR_RATING,
            measure_id = str_c(measure_id, "_STAR_RATING")) %>%
  pivot_wider(id_cols = c(cc_date, provider_id),
              names_from = measure_id, values_from = STAR_RATING)

hcahps_denom_df <-
  temp_sas_hcahps_df %>%
  filter(!(measure_id %in% c("H_INDI","H_GLOB"))) %>%
  select(cc_date, provider_id, H_NUMB_COMP, H_RESP_RATE_P) %>%
  distinct()

sas_hcahps_df <-
  full_join(hcahps_star_df, hcahps_denom_df, by = c("cc_date", "provider_id"))
```


# Combine Columns

```{r}
sas_input_file_hospital_ids <-
  all_hospital_info_df %>%
  filter(cc_date == "2023-01-01", in_sas_input_file) %>%
  pull(provider_id)

sas_df <-
  sas_most_groups_df %>%
  full_join(sas_hai_df, by = c("cc_date", "provider_id")) %>%
  full_join(sas_hcahps_df, by = c("cc_date", "provider_id")) %>%
  rename(PROVIDER_ID = provider_id) %>%
  filter(PROVIDER_ID %in% sas_input_file_hospital_ids)

my_sas_input_df <-
  sas_df %>%
  group_by(cc_date) %>%
  nest()

my_sas_input_df
my_sas_input_df$data[[25]]
```


```{r}
save(my_sas_input_df, file = "rdata/my_sas_input_df.RData")
```







## Note about the 2023-01 SAS input

When scores are reported in both the 2023-01 Care Compare report and in the SAS input file released July '23, the scores are the same. However, there are about 7 Process measures (IMM_3,OP_22,OP_13,OP_29,SEP_1,OP18B,OP_10) where a significant number of hospitals have a score that's either (a) reported only in the Care Compare data or (b) reported only in the SAS sample input file.

There wasn't a clear pattern explaining why scores may be reported in one source but not the other. For all 7 measures, the problem above occurred in no more than 6% of the hospitals that had a score reported in at least one of the sources. So there didn't seem to be other discrepancies between these two data sources.

To avoid confusion in reporting, I replaced the SAS input constructed from Care Compare reports with the SAS sample input. Then all computations were consistent with CMS's preview reports.



```{r}
install.packages("devtools")
devtools::install_github("jimurick/hospitalstars")
```



```{r}
library(hospitalstars)

to_replace   <- which(my_sas_input_df$cc_date == "2023-01-01")
replace_with <- which(all_sas_input_df$cc_date == "2023-01-01")

my_sas_input2_df <- my_sas_input_df
sas_replacement_df <- all_sas_input_df$data[[replace_with]] %>%
  rename(
    PSI_04 = PSI_4_SURG_COMP, PSI_90 = PSI_90_SAFETY,
    PSI_04_DEN = PSI_4_SURG_COMP_DEN, PSI_90_DEN = PSI_90_SAFETY_DEN
  ) %>%
  select_at(colnames(my_sas_input2_df$data[[to_replace]]))
my_sas_input2_df$data[[to_replace]] <- sas_replacement_df

sas_package_data_df <-
  my_sas_input2_df %>%
  mutate(sas_output_df = map(data, hospitalstars::compute_star_scores)) %>%
  select(-data)
```






```{r}
save(sas_package_data_df, file = "rdata/sas_package_data_df.RData")
```
