==> Stars - Hospitals/05_sas_data_for_package.Rmd <==
---
title: "R Notebook"
output:
html_notebook: default
---



```{r}
source("init.R")
library(data.table)

load(file = file.path(dirs$rdata, "my_sas_input_df.RData"))
load(file = file.path(dirs$rdata, "all_data_list.RData"))
load(file = file.path(dirs$rdata, "measures_202104_dt.rda"))
load(file = file.path(dirs$rdata, "measure_groups_202104_dt.rda"))
#load(file = file.path(dirs$rdata, "measure_ref.rda"))

all_mort_readm_data_df <- all_data_list$all_mort_readm_data_df
all_maternal_data_df   <- all_data_list$all_maternal_data_df
all_hai_data_df        <- all_data_list$all_hai_data_df
all_hcahps_data_df     <- all_data_list$all_hcahps_data_df
all_timely_data_df     <- all_data_list$all_timely_data_df
all_imaging_data_df    <- all_data_list$all_imaging_data_df
all_ipfqr_data_df      <- all_data_list$all_ipfqr_data_df
all_hospital_info_df   <- all_data_list$all_hospital_info_df
all_measure_info_df    <- all_data_list$all_measure_info_df
all_va_hospital_df     <- all_data_list$all_va_hospital_df
all_sas_input_df       <- all_data_list$all_sas_input_df
all_national_rates_df  <- all_data_list$all_national_rates_df
measure_id_history_df  <- all_data_list$measure_id_history_df
```





```{r}
measures_202104_dt <-
  measures_202104_dt %>%
  mutate(
    measure_std = case_when(
      measure_id == "PSI_4_SURG_COMP" ~ "std_PSI_04",
      measure_id == "PSI_90_SAFETY"   ~ "std_PSI_90",
      TRUE ~ measure_std
    ),
    measure_id = case_when(
      measure_id == "PSI_4_SURG_COMP" ~ "PSI_04",
      measure_id == "PSI_90_SAFETY"   ~ "PSI_90",
      TRUE ~ measure_id
    )
  ) %>%
  as.data.table()

measures_202307_dt <-
  data.table::rbindlist(list(
    measures_202104_dt[
      !(measures_202104_dt$retired_2022), !c("retired_2021", "retired_2022")
    ],
    all_measure_info_df %>%
      filter(measure_id == "HCP_COVID_19", cc_date == "2023-07-01") %>%
      transmute(
        measure_id,
        group_name   = measure_group,
        measure_desc = measure_name,
        flip_sign    = ifelse(higher_is_better == TRUE, 1, -1),
        measure_std  = measure_id
      )
  ))

measure_ref <-
  list(
    "group_dt" = measure_groups_202104_dt,
    "measure_dt" = list(
      "v202104" = measures_202104_dt[, !c("retired_2021", "retired_2022")],
      "v202207" = measures_202104_dt[!measures_202104_dt$retired_2021, !c("retired_2021", "retired_2022")],
      "v202307" = measures_202307_dt
    )
  )
```


```{r}
save(measure_ref, file = "rdata/measure_ref.rda")
```



```{r}
more_measures_info_dt <-
  measures_202104_dt %>%
  mutate(
    measure_std = case_when(
      measure_id == "PSI_4_SURG_COMP" ~ "std_PSI_04",
      measure_id == "PSI_90_SAFETY"   ~ "std_PSI_90",
      TRUE ~ measure_std
    ),
    measure_id = case_when(
      measure_id == "PSI_4_SURG_COMP" ~ "PSI_04",
      measure_id == "PSI_90_SAFETY"   ~ "PSI_90",
      TRUE ~ measure_id
    )
  )
# 
# measures_202104_dt <-
#   measures_202104_dt %>%
#   mutate(
#     measure_id = case_when(
#       measure_id == "PSI_4_SURG_COMP" ~ "PSI_04",
#       measure_id == "PSI_90_SAFETY"   ~ "PSI_90",
#       TRUE ~ measure_id
#     ),
#     measure_std = case_when(
#       measure_std == "std_PSI_4_SURG_COMP" ~ "std_PSI_04",
#       measure_std == "std_PSI_90_SAFETY"   ~ "std_PSI_90",
#       TRUE ~ measure_id
#     )
#   ) %>%
#   data.table::as.data.table()
# 
# measure_ref_202104 <-
#   list(
#     "measure_dt" = measures_202104_dt,
#     "group_dt" = measure_groups_202104_dt
#   )

```


```{r}
star_algorithm_input_list <- setNames(
    lapply(my_sas_input_df$data, as.data.frame),
    str_sub(as.character(my_sas_input_df$cc_date), end = 7)
  )

star_algorithm_input_df <-
  my_sas_input_df %>%
  unnest(data) %>%
  rename(report_date = cc_date) %>%
  as.data.frame()


hospital_results_df <-
  all_hospital_info_df %>%
  filter(cc_date >= "2018-01-01") %>%
  transmute(
    report_month = map_chr(cc_date, ~strftime(.x, "%Y%m")),
    provider_id,
    peer_group = measure_groups,
    star_rating = overall_rating
  ) %>%
  pivot_longer(peer_group:star_rating, names_to = "x", values_to = "y") %>%
  filter(!is.na(y)) %>%
  transmute(
    provider_id,
    x = str_c(x, "_", report_month),
    y
  ) %>%
  pivot_wider(provider_id, names_from = x, values_from = y)

hospital_info_df <-
  all_data_list$all_hospital_info_df %>%
  filter(cc_date >= "2018-01-01") %>%
  select(cc_date, provider_id, hospital_name, hospital_type,
                address, city, state, zip, county, phone) %>%
  group_by(provider_id) %>%
  nest() %>%
  mutate(data = map(data, ~arrange(.x, desc(cc_date)) %>% slice(1))) %>%
  unnest(data) %>%
  select(-cc_date) %>%
  left_join(hospital_results_df, by = "provider_id") %>%
  rename(PROVIDER_ID = provider_id) %>%
  as.data.frame()

hospital_info_df
```


```{r}
measure_info_df <-
  all_data_list$all_measure_info_df %>%
  filter(cc_date >= "2018-01-01", !is.na(care_compare_filename)) %>%
  #filter(!is.na(sas_measure_id)) %>%
  mutate(
    measure_id = case_when(
      measure_id == "PSI_4_SURG_COMP" ~ "PSI_04",
      measure_id == "PSI_90_SAFETY" ~ "PSI_90",
      measure_id == "IMM_3_OP_27" ~ "IMM_3",
      TRUE ~ measure_id
    ),
    sas_measure_id = case_when(
      sas_measure_id == "PSI_4_SURG_COMP" ~ "PSI_04",
      sas_measure_id == "PSI_90_SAFETY" ~ "PSI_90",
      sas_measure_id == "IMM_3_OP_27" ~ "IMM_3",
      grepl("LINEAR$", sas_measure_id) ~ gsub("LINEAR$", "STAR_RATING", sas_measure_id),
      TRUE ~ sas_measure_id
    )
  ) %>%
  transmute(
    report_date = cc_date,
    measure_group,
    measure_id = ifelse(is.na(sas_measure_id), measure_id, sas_measure_id),
    measure_name = case_when(
      measure_id == "H_CLEAN_STAR_RATING"  ~ "Cleanliness",
      measure_id == "H_COMP_1_STAR_RATING" ~ "Nurse communication",
      measure_id == "H_COMP_2_STAR_RATING" ~ "Doctor communication",
      measure_id == "H_COMP_3_STAR_RATING" ~ "Staff responsiveness",
      measure_id == "H_COMP_5_STAR_RATING" ~ "Communication about medicines",
      measure_id == "H_COMP_6_STAR_RATING" ~ "Discharge information",
      measure_id == "H_COMP_7_STAR_RATING" ~ "Care transition",
      measure_id == "H_HSP_RATING_STAR_RATING" ~ "Overall hospital rating",
      measure_id == "H_QUIET_STAR_RATING"  ~ "Quietness",
      measure_id == "H_RECMND_STAR_RATING" ~ "Recommend hospital",
      TRUE ~ measure_name
    ),
    higher_is_better,
    data_start_date = start_date, data_end_date = end_date,
    data_start_quarter = start_quarter, data_end_quarter = end_quarter,
    source_filename = care_compare_filename
  ) %>%
  as.data.frame()


two_weird_hcahps_measures <-
  more_measures_info_dt %>%
  as.tibble() %>%
  filter(measure_id %in% c("H_GLOB_STAR_RATING","H_INDI_STAR_RATING")) %>%
  transmute(measure_id, measure_name = measure_desc)

measure_info_to_add_df <-
  measure_info_df %>%
  filter(report_date < "2020-10-01", measure_group == "Patient Experience") %>%
  distinct(
    report_date, measure_group, higher_is_better,
    data_start_date, data_end_date, data_start_quarter, data_end_quarter
  ) %>%
  mutate(
    data = map(report_date, function(x) two_weird_hcahps_measures),
    source_filename = NA_character_
  ) %>%
  unnest(data)

measure_info_df <-
  bind_rows(measure_info_df, measure_info_to_add_df) %>%
  arrange(report_date, measure_group, measure_id) %>%
  as.data.frame()

measure_info_df %>% arrange(desc(report_date))
```


```{r}
sas_package_input_v202104_df <-
  haven::read_sas(file.path(dirs$sas, "all_data_2021apr.sas7bdat"))
sas_package_input_v202207_df <-
  haven::read_sas(file.path(dirs$sas, "all_data_2022jul.sas7bdat"))
sas_package_input_v202307_df <-
  haven::read_sas(file.path(dirs$sas, "alldata_2023jul.sas7bdat"))
sas_package_output_v202104_df <-
	read.csv(file.path(dirs$sas, "sas_package_output_v202104_202010.csv"),
	         colClasses = c("PROVIDER_ID"="character"))
sas_package_output_v202207_df <-
	read.csv(file.path(dirs$sas, "sas_package_output_v202207_202107.csv"),
	         colClasses = c("PROVIDER_ID"="character"))
sas_package_output_v202307_df <-
	read.csv(file.path(dirs$sas, "sas_package_output_v202307_202301.csv"),
	         colClasses = c("PROVIDER_ID"="character"))
fix_psi_cols <- function(df) {
  colnames(df) <-
    gsub("PSI_90_SAFETY", "PSI_90",
         gsub("PSI_4_SURG_COMP", "PSI_04", colnames(df)))
  df %>% as.data.frame()
}
sas_package_input_v202104_df  <- fix_psi_cols(sas_package_input_v202104_df)
sas_package_input_v202207_df  <- fix_psi_cols(sas_package_input_v202207_df)
sas_package_input_v202307_df  <- fix_psi_cols(sas_package_input_v202307_df)
sas_package_output_v202104_df <- fix_psi_cols(sas_package_output_v202104_df)
sas_package_output_v202207_df <- fix_psi_cols(sas_package_output_v202207_df)
sas_package_output_v202307_df <- fix_psi_cols(sas_package_output_v202307_df)

sas_package <- list(
	"v202104" = list(
	  "input_df"=sas_package_input_v202104_df,
	  "output_df"=sas_package_output_v202104_df
	),
	"v202207" = list(
	  "input_df"=sas_package_input_v202207_df,
	  "output_df"=sas_package_output_v202207_df
	),
	"v202307" = list(
	  "input_df"=sas_package_input_v202307_df,
	  "output_df"=sas_package_output_v202307_df
	)
)
```




```{r}
hospitalstars_data <- list(
  "hospital_info_df" = hospital_info_df,
  "measure_info_df" = measure_info_df,
  "star_algorithm_input_df" = star_algorithm_input_df,
  #"measure_ref" = measure_ref_202104,
  "measure_ref" = measure_ref,
  "sas_package" = sas_package
)
save(hospitalstars_data, file = file.path(dirs$rdata, "hospitalstars_data.RData"))
```
